Extract DNA mtDNA helpers
