+++ service-worker.js
@@ -239,6 +239,7 @@ const APP_SHELL = [
   '/js/dna-actions.js',
   '/js/dna-genotype.js',
   '/js/dna.js',
+  '/js/dna-mtdna.js',
   '/js/hardware.js',
   '/js/sync.js',
   '/js/sync-configure.js',
